requirements.txt
fastapi
uvicorn

psycopg2-binary
sqlalchemy
pydantic

pytest
pytest-asyncio
coverage
flake8

python-dotenv
httpx
